# internal/handlers/console/openapi.yaml
openapi: 3.1.0
info:
  title: 听见 Console API
  version: 0.1.0
  description: |
    Stage B 控制台后端契约。所有路径前缀 `/fb/v1/console`。
    单一真理源——任何 endpoint 的 schema/参数/错误一切以本文件为准。
    前端 TS 类型由 `openapi-typescript` 自动生成到
    `listen/console/src/api/types.ts`，CI gate 阻止漂移。

    认证模型：HMAC-signed HttpOnly cookie + 双提交 CSRF header。
    所有非 GET/HEAD 请求必须带 `X-CSRF-Token`。
servers:
  - url: https://listen.app
    description: 生产环境
  - url: http://127.0.0.1:8090
    description: 本地开发
...
paths:
...
  /fb/v1/console/install/callback:
    get:
      tags: [install]
      summary: 飞书 OAuth 回调
      description: |
        飞书 redirects here 携带 `code` 与 `state`。
        服务端用 code 换 user_access_token，取 tenant_key + user info，
        upsert tenants 行，签 session cookie，302 回 `/console`。
      parameters:
        - in: query
          name: code
          required: true
          schema: { type: string }
        - in: query
          name: state
          required: true
          schema: { type: string }
      responses:
        '302':
          description: 登录成功，重定向到 console 主页
          headers:
            Set-Cookie:
              schema: { type: string }
              description: listen_session（HttpOnly, Secure, SameSite=Lax）
        '400':
          $ref: '#/components/responses/BadRequest'
...
  /fb/v1/console/api-keys:
    get:
      tags: [api-keys]
      summary: 列 API key（仅返回元数据，不返回明文）
      security:
        - sessionCookie: []
      responses:
        '200':
          description: API key 列表
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/ApiKey' }

    post:
      tags: [api-keys]
      summary: 铸新 key（明文 secret 仅此一次返回）
      description: |
        明文 key 仅本响应里出现，之后只能查到 prefix。
        客户必须当场保存——丢失只能撤销重铸。
      security:
        - sessionCookie: []
        - csrfHeader: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [label]
              properties:
                label:
                  type: string
                  maxLength: 200
                  description: 用途备注，如 "prod-app" / "dev-test"
      responses:
        '201':
          description: 新建成功
          content:
            application/json:
              schema: { $ref: '#/components/schemas/NewApiKey' }
        '400':
          $ref: '#/components/responses/BadRequest'
  /fb/v1/console/api-keys/{id}:
    delete:
      tags: [api-keys]
      summary: 撤销 key（软删，留审计记录）
      security:
        - sessionCookie: []
        - csrfHeader: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      responses:
        '204':
          description: 已撤销
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: listen_session
    csrfHeader:
      type: apiKey
      in: header
      name: X-CSRF-Token
  schemas:
...
    ApiKey:
      type: object
      required: [id, key_prefix, label, is_active, created_at]
      properties:
        id: { type: string, format: uuid }
        key_prefix:
          type: string
          description: 前缀 + ... + 后 4 位，用于客户辨识
        label: { type: string }
        is_active: { type: boolean }
        created_at: { type: string, format: date-time }
        last_used_at:
          type: string
          format: date-time
          nullable: true
        revoked_at:
          type: string
          format: date-time
          nullable: true
    NewApiKey:
      allOf:
        - $ref: '#/components/schemas/ApiKey'
        - type: object
          required: [secret]
          properties:
            secret:
              type: string
              description: 明文 key。仅此一次返回，丢失只能撤销重铸。
...
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: 机器可读，如 `unauthorized` / `tenant_not_found`
        message:
          type: string
          description: 中文人类可读
        request_id:
          type: string
          description: trace id，复用 chi RequestID 中间件
  responses:
    BadRequest:
      description: 请求体或参数不合法
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
    NotFound:
      description: 资源不存在或不属于当前 tenant
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...